[RFC] vbdev_redirector: Add Adaptive Distributed NVMe-oF Namespaces
ADNN adapts NVMe-oF to distributed block storage systems. When a
volume is scattered over several storage nodes, point to point
protocols like NVMe-oF can't get host IO directly to every LBA of
that volume. These systems typically use custom block layer clients
in the hosts, or route IO to a gateway that forwards it to the right
storage node. ADNN extends NVMe-oF by exposing these volumes from
NVMe-oF targets in all the storage nodes, and enabling hosts to
learn which volume extents are located on which targets.

See module/bdev/redirector/README for more details. Section 6 of the
README may be helpful when reviewing this patch or navigating the
code.

Signed-off-by: Scott Peterson <scott.d.peterson@intel.com>
Change-Id: I2ed865eafa9c69e80538b10a5850331d563772be