[RFC] vbdev_split: add "share" option
Share option to vbdev_split makes n ports to the base bdev rather than
n partitions.

The split vbdev does its thing with partitions, so the shares are of
course still bdev partitions but they all access the entire LBA range
of the base.

This is needed for some tests of bdevs we expect to allow shared access
to. Normally a bdev can only be claimed by one opener, and a claim is
required to write to a bdev.

Allow NVME_ADMIN to pass through partitions that are the same size as
their base bdev

Signed-off-by: Scott Peterson <scott.d.peterson@intel.com>
Change-Id: Ie44e72182efa54d491bfb136e4823faa76b945ba